added findCardIndex function
Added findCardIndex(Card c) for searching for a card in a player's hand to get the index of the first instance of that card. I did this because the MouseAdapter can return a Component(which i downcast to CardPane), but there's no way for me to know where in the array that cardpane is.