Update Uno.java
Converted Uno.java to a class that extends JFrame and started implementing GUI elements as well as working in the game logic into the event handler. 
It works mostly, skip and reverse should still work fine. Just click the card in the frame to play that card. 
I know I broke the multiple draw 2 block for the moment, specifically at line 357 where it was taking input from the Scanner. 
Also commented out Wild card functionality until we add GUI components for that. 
I was hoping to get this to complete working order with all your game logic but I ran into a lot of snags figuring out the correct way to implement the UI. I wanted to get the base of it out to continue another time.